+++ src/ui/main.ts
@@ -20,7 +20,6 @@ import { Indicators, toggleManualIndicatorLit } from "./custom/indicators"
 import WarningWindow from "./custom/warning"
 import DynamicDashboard from "../common/plugin"
 import MathUtils from "../utils/math_utils"
-import { ParkAndScenarioData } from "../data/park_and_scenario"
 
 /**
  * Whether the window is open.
@@ -568,6 +567,7 @@ function menu(): void {
               groupbox({
                 text: language.ui.main.groupbox.park_and_scenario.park_rating,
                 content: [
+                  // Park Rating Current
                   horizontal({
                     content: [
                       label({
@@ -605,8 +605,8 @@ function menu(): void {
                           (value) => {
                             return MathUtils.normalise(
                               value,
-                              ParkAndScenarioData.MIN_PARK_RATING,
-                              ParkAndScenarioData.MAX_PARK_RATING
+                              Data.ParkAndScenarioData.MIN_PARK_RATING,
+                              Data.ParkAndScenarioData.MAX_PARK_RATING
                             )
                           }
                         ),
@@ -615,6 +615,7 @@ function menu(): void {
                       })
                     ]
                   }),
+                  // Park Rating Average
                   horizontal({
                     content: [
                       label({
@@ -653,8 +654,8 @@ function menu(): void {
                           (value) => {
                             return MathUtils.normalise(
                               value,
-                              ParkAndScenarioData.MIN_PARK_RATING,
-                              ParkAndScenarioData.MAX_PARK_RATING
+                              Data.ParkAndScenarioData.MIN_PARK_RATING,
+                              Data.ParkAndScenarioData.MAX_PARK_RATING
                             )
                           }
                         ),
@@ -663,6 +664,7 @@ function menu(): void {
                       })
                     ]
                   }),
+                  // Park Rating Year Average
                   horizontal({
                     content: [
                       label({
@@ -702,8 +704,8 @@ function menu(): void {
                           (value) => {
                             return MathUtils.normalise(
                               value,
-                              ParkAndScenarioData.MIN_PARK_RATING,
-                              ParkAndScenarioData.MAX_PARK_RATING
+                              Data.ParkAndScenarioData.MIN_PARK_RATING,
+                              Data.ParkAndScenarioData.MAX_PARK_RATING
                             )
                           }
                         ),
@@ -712,6 +714,7 @@ function menu(): void {
                       })
                     ]
                   }),
+                  // Park Rating Month Average
                   horizontal({
                     content: [
                       label({
@@ -751,13 +754,76 @@ function menu(): void {
                           (value) => {
                             return MathUtils.normalise(
                               value,
-                              ParkAndScenarioData.MIN_PARK_RATING,
-                              ParkAndScenarioData.MAX_PARK_RATING
+                              Data.ParkAndScenarioData.MIN_PARK_RATING,
+                              Data.ParkAndScenarioData.MAX_PARK_RATING
                             )
                           }
                         ),
                         background: Colour.Grey,
                         foreground: Colour.BrightGreen
+                      })
+                    ]
+                  }),
+                  // Park Rating Warning Days
+                  horizontal({
+                    content: [
+                      label({
+                        text: compute(
+                          baseData.local.park_and_scenario
+                            .park_rating_warning_days.store,
+                          baseData.local.options.display_mode.store,
+                          (value, mode) => {
+                            switch (mode) {
+                              case Data.Options.DisplayMode.PROGRESS_BAR:
+                                return language.ui.main.label
+                                  .park_rating_warning_days
+                              case Data.Options.DisplayMode.VALUE:
+                                return tr(
+                                  language.ui.main.label
+                                    .park_rating_warning_days,
+                                  Data.ParkAndScenarioData
+                                    .MAX_RATING_WARNING_DAYS - value
+                                )
+                              default:
+                                return ""
+                            }
+                          }
+                        ),
+                        tooltip:
+                          language.ui.main.tooltip.park_rating_warning_days
+                      }),
+                      progressBar({
+                        visibility: compute(
+                          baseData.local.options.display_mode.store,
+                          (value) => {
+                            return value ===
+                              Data.Options.DisplayMode.PROGRESS_BAR
+                              ? "visible"
+                              : "none"
+                          }
+                        ),
+                        percentFilled: compute(
+                          baseData.local.park_and_scenario
+                            .park_rating_warning_days.store,
+                          () => {
+                            return Data.ParkAndScenarioData.getWarningDaysPercentage()
+                          }
+                        ),
+                        background: Colour.Grey,
+                        foreground: compute(
+                          baseData.local.park_and_scenario
+                            .park_rating_warning_days.store,
+                          () => {
+                            if (
+                              Data.ParkAndScenarioData.getWarningDaysPercentage() <=
+                              Data.ParkAndScenarioData
+                                .RATING_WARNING_DAYS_THRESHOLD
+                            ) {
+                              return Colour.BrightRed
+                            }
+                            return Colour.BrightGreen
+                          }
+                        )
                       })
                     ]
                   })